pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tradeiros-python"
version = "0.1.0"
description = "Uma biblioteca Python."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "python-dotenv",
    "python-okx >= 0.4.1",
    "ccxt"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
